reorg backend code
